refactor: Add compiler.ts (exports evalPSVG/compilePSVG functions)
