+++ src/psvg.ts
@@ -1,19 +1,8 @@
-import { parsePSVG } from './parser';
-import { PSVGFunc, transpilePSVG } from './transpiler';
+import { evalPSVG, compilePSVG } from './compiler';
 
-export { parsePSVG, PSVGFunc, transpilePSVG };
-
-export function evalPSVG(js: string): string {
-  return Function(`"use strict";${js};return __out;`)();
-}
-
-export function compilePSVG(psvg: string): string {
-  let prgm = parsePSVG(psvg);
-  // console.dir(prgm,{depth:null});
-  let js = transpilePSVG(prgm);
-  // console.log(js);
-  return evalPSVG(js);
-}
+export { parsePSVG } from './parser';
+export { PSVGFunc, transpilePSVG } from './transpiler';
+export { evalPSVG, compilePSVG };
 
 if (typeof window !== 'undefined') {
   window.addEventListener('load', function () {
